+++ api-examples/eager-tagger.py
@@ -1,16 +1,29 @@
 
+import os
 from eight_mile.utils import listify
 import baseline
 import eight_mile.tf.layers as L
-
+from eight_mile.utils import get_version
 import eight_mile.embeddings
 from eight_mile.tf.embeddings import LookupTableEmbeddings
-from  eight_mile.w2v import PretrainedEmbeddingsModel, RandomInitVecModel
+from eight_mile.w2v import PretrainedEmbeddingsModel, RandomInitVecModel
 from eight_mile.tf.optz import EagerOptimizer
+from baseline.tf.tfy import SET_TRAIN_FLAG
 import tensorflow as tf
 import logging
 import numpy as np
-tf.enable_eager_execution()
+
+
+TF_VERSION = get_version(tf)
+if TF_VERSION < 2:
+    from tensorflow import count_nonzero
+    tf.enable_eager_execution()
+    Optimizer = tf.train.GradientDescentOptimizer
+
+else:
+    from tensorflow.compat.v1 import count_nonzero
+    Optimizer = tf.optimizers.SGD
+
 
 NUM_PREFETCH = 2
 SHUF_BUF_SZ = 5000
@@ -43,11 +56,11 @@ def to_tensors(ts):
     return np.stack(X), np.stack(Xch), np.stack(y)
 
 
-VSM_MODEL = '/home/dpressel/.bl-data/dce69c404025a8312c323197347695e81fd529fc/glove.twitter.27B.200d.txt'
+VSM_MODEL = os.path.expanduser('~/.bl-data/dce69c404025a8312c323197347695e81fd529fc/glove.twitter.27B.200d.txt')
 
-TS = '/home/dpressel/dev/work/baseline/data/oct27.train'
-VS = '/home/dpressel/dev/work/baseline/data/oct27.dev'
-ES = '/home/dpressel/dev/work/baseline/data/oct27.test'
+TS = os.path.expanduser('~/dev/work/baseline/data/oct27.train')
+VS = os.path.expanduser('~/dev/work/baseline/data/oct27.dev')
+ES = os.path.expanduser('~/dev/work/baseline/data/oct27.test')
 
 
 feature_desc = {
@@ -92,37 +105,39 @@ X_test, Xch_test, y_test = to_tensors(reader.load(test_file, vocabs=vocabs, batc
 
 
 def train_input_fn():
+    SET_TRAIN_FLAG(True)
     dataset = tf.data.Dataset.from_tensor_slices((X_train, Xch_train, y_train))
     dataset = dataset.shuffle(buffer_size=SHUF_BUF_SZ)
     # https://github.com/tensorflow/tensorflow/blob/master/tensorflow/contrib/distribute/README.md
     # effective_batch_sz = args.batchsz*args.gpus
     dataset = dataset.batch(20)
     #dataset = dataset.map(lambda x, y: (x, y))
-    dataset = dataset.map(lambda x, xch, y: ({'word': x, 'char': xch, 'lengths': tf.count_nonzero(x, axis=1)}, y))
+    dataset = dataset.map(lambda x, xch, y: ({'word': x, 'char': xch, 'lengths': count_nonzero(x, axis=1)}, y))
     dataset = dataset.repeat(1)
     dataset = dataset.prefetch(NUM_PREFETCH)
-    _ = dataset.make_one_shot_iterator()
+    #_ = dataset.make_one_shot_iterator()
     return dataset
 
 
 def eval_input_fn():
+    SET_TRAIN_FLAG(False)
     dataset = tf.data.Dataset.from_tensor_slices((X_valid, Xch_valid, y_valid))
     dataset = dataset.batch(20)
-    dataset = dataset.map(lambda x, xch, y: ({'word': x, 'char': xch, 'lengths': tf.count_nonzero(x, axis=1)}, y))
+    dataset = dataset.map(lambda x, xch, y: ({'word': x, 'char': xch, 'lengths': count_nonzero(x, axis=1)}, y))
     #dataset = dataset.map(lambda x, y: (x, y))
-    _ = dataset.make_one_shot_iterator()
+    #_ = dataset.make_one_shot_iterator()
     return dataset
 
 
 def predict_input_fn():
+    SET_TRAIN_FLAG(False)
     dataset = tf.data.Dataset.from_tensor_slices((X_test, y_test))
     dataset = dataset.batch(1)
-    dataset = dataset.map(lambda x, xch, y: ({'word': x, 'char': xch, 'lengths': tf.count_nonzero(x, axis=1)}, y))
-    _ = dataset.make_one_shot_iterator()
+    dataset = dataset.map(lambda x, xch, y: ({'word': x, 'char': xch, 'lengths': count_nonzero(x, axis=1)}, y))
     return dataset
 
 
-transducer = L.BiLSTMEncoder(200, 1, 0.5, output_fn=L.rnn_signal)
+transducer = L.BiLSTMEncoderSequence(None, 200, 2, 0.5)
 model = L.TagSequenceModel(len(labels), embeddings, transducer)
 
 train_loss_results = []
@@ -134,7 +149,7 @@ def loss(model, x, y):
   return model.decoder_model.neg_log_loss(unary, y, x['lengths'])
 
 
-optim = EagerOptimizer(loss, tf.train.GradientDescentOptimizer(learning_rate=0.01))
+optim = EagerOptimizer(loss, Optimizer(learning_rate=0.01))
 
 
 import time
@@ -152,8 +167,8 @@ for epoch in range(num_epochs):
 
         loss_acc += loss_value
         step += 1
-    print('training time {}'.format(time.time() - start))
 
+    print('training time {}'.format(time.time() - start))
     mean_loss = loss_acc / step
     print('Training Loss {}'.format(mean_loss))
