+++ python/baseline/tf/classify/training/feed.py
@@ -1,0 +1,104 @@
+import six
+import os
+import time
+import logging
+import tensorflow as tf
+
+from eight_mile.confusion import ConfusionMatrix
+from eight_mile.progress import create_progress_bar
+from eight_mile.utils import listify
+from eight_mile.tf.optz import optimizer
+
+from baseline.utils import get_model_file, get_metric_cmp
+from baseline.tf.tfy import _add_ema, TRAIN_FLAG, SET_TRAIN_FLAG
+from baseline.tf.classify.training.utils import to_tensors
+from baseline.train import EpochReportingTrainer, create_trainer, register_trainer, register_training_func
+from baseline.utils import verbose_output
+from baseline.model import create_model_for
+import numpy as np
+
+
+@register_training_func('classify', 'feed_dict')
+def fit(model_params, ts, vs, es=None, **kwargs):
+    """
+    Train a classifier using TensorFlow with a `feed_dict`.  This
+    is the previous default behavior for training.  To use this, you need to pass
+    `fit_func: feed_dict` in your MEAD config
+
+    :param model_params: The model (or parameters to create the model) to train
+    :param ts: A training data set
+    :param vs: A validation data set
+    :param es: A test data set, can be None
+    :param kwargs:
+        See below
+
+    :Keyword Arguments:
+        * *do_early_stopping* (``bool``) --
+          Stop after evaluation data is no longer improving.  Defaults to True
+        * *verbose* (`dict`) A dictionary containing `console` boolean and `file` name if on
+        * *epochs* (``int``) -- how many epochs.  Default to 20
+        * *outfile* -- Model output file, defaults to classifier-model.pyth
+        * *patience* --
+           How many epochs where evaluation is no longer improving before we give up
+        * *reporting* --
+           Callbacks which may be used on reporting updates
+        * *nsteps* (`int`) -- If we should report every n-steps, this should be passed
+        * *ema_decay* (`float`) -- If we are doing an exponential moving average, what decay to us4e
+        * *clip* (`int`) -- If we are doing gradient clipping, what value to use
+        * *optim* (`str`) -- The name of the optimizer we are using
+        * *lr* (`float`) -- The learning rate we are using
+        * *mom* (`float`) -- If we are using SGD, what value to use for momentum
+        * *beta1* (`float`) -- Adam-specific hyper-param, defaults to `0.9`
+        * *beta2* (`float`) -- Adam-specific hyper-param, defaults to `0.999`
+        * *epsilon* (`float`) -- Adam-specific hyper-param, defaults to `1e-8
+
+    :return: None
+    """
+    do_early_stopping = bool(kwargs.get('do_early_stopping', True))
+    verbose = kwargs.get('verbose', {'console': kwargs.get('verbose_console', False), 'file': kwargs.get('verbose_file', None)})
+    epochs = int(kwargs.get('epochs', 20))
+    model_file = get_model_file('classify', 'tf', kwargs.get('basedir'))
+    ema = True if kwargs.get('ema_decay') is not None else False
+
+    best_metric = 0
+    if do_early_stopping:
+        early_stopping_metric = kwargs.get('early_stopping_metric', 'acc')
+        early_stopping_cmp, best_metric = get_metric_cmp(early_stopping_metric, kwargs.get('early_stopping_cmp'))
+        patience = kwargs.get('patience', epochs)
+        print('Doing early stopping on [%s] with patience [%d]' % (early_stopping_metric, patience))
+
+    reporting_fns = listify(kwargs.get('reporting', []))
+    print('reporting', reporting_fns)
+
+    TRAIN_FLAG()
+    trainer = create_trainer(model_params, **kwargs)
+
+    last_improved = 0
+
+    for epoch in range(epochs):
+
+        trainer.train(ts, reporting_fns, dataset=False)
+        test_metrics = trainer.test(vs, reporting_fns, phase='Valid', dataset=False)
+
+        if do_early_stopping is False:
+            trainer.checkpoint()
+            trainer.model.save(model_file)
+
+        elif early_stopping_cmp(test_metrics[early_stopping_metric], best_metric):
+            last_improved = epoch
+            best_metric = test_metrics[early_stopping_metric]
+            print('New best %.3f' % best_metric)
+            trainer.checkpoint()
+            trainer.model.save(model_file)
+
+        elif (epoch - last_improved) > patience:
+            print('Stopping due to persistent failures to improve')
+            break
+
+    if do_early_stopping is True:
+        print('Best performance on %s: %.3f at epoch %d' % (early_stopping_metric, best_metric, last_improved))
+
+    if es is not None:
+        print('Reloading best checkpoint')
+        trainer.recover_last_checkpoint()
+        trainer.test(es, reporting_fns, phase='Test', verbose=verbose, dataset=False)
