+++ python/baseline/tf/tagger/training/feed.py
@@ -1,0 +1,108 @@
+import six
+import os
+import time
+import logging
+from eight_mile.utils import listify
+
+from baseline.train import create_trainer, register_training_func
+from baseline.tf.tfy import TRAIN_FLAG
+from baseline.utils import get_model_file, get_metric_cmp
+from baseline.tf.tfy import reload_lower_layers
+from baseline.tf.tagger.training.utils import TaggerEvaluatorTf
+
+logger = logging.getLogger('baseline')
+
+
+@register_training_func('tagger', 'feed_dict')
+def fit(model_params, ts, vs, es, **kwargs):
+    """
+    Train a classifier using TensorFlow with a `feed_dict`.  This
+    is the previous default behavior for training.  To use this, you need to pass
+    `fit_func: feed_dict` in your MEAD config
+
+    :param model_params: The model to train
+    :param ts: A training data set
+    :param vs: A validation data set
+    :param es: A test data set, can be None
+    :param kwargs:
+        See below
+
+    :Keyword Arguments:
+        * *do_early_stopping* (``bool``) --
+          Stop after evaluation data is no longer improving.  Defaults to True
+        * *verbose* (`dict`) A dictionary containing `console` boolean and `file` name if on
+        * *epochs* (``int``) -- how many epochs.  Default to 20
+        * *outfile* -- Model output file, defaults to classifier-model.pyth
+        * *patience* --
+           How many epochs where evaluation is no longer improving before we give up
+        * *reporting* --
+           Callbacks which may be used on reporting updates
+        * *nsteps* (`int`) -- If we should report every n-steps, this should be passed
+        * *ema_decay* (`float`) -- If we are doing an exponential moving average, what decay to us4e
+        * *clip* (`int`) -- If we are doing gradient clipping, what value to use
+        * *optim* (`str`) -- The name of the optimizer we are using
+        * *lr* (`float`) -- The learning rate we are using
+        * *mom* (`float`) -- If we are using SGD, what value to use for momentum
+        * *beta1* (`float`) -- Adam-specific hyper-param, defaults to `0.9`
+        * *beta2* (`float`) -- Adam-specific hyper-param, defaults to `0.999`
+        * *epsilon* (`float`) -- Adam-specific hyper-param, defaults to `1e-8
+
+    :return: None
+    """
+    epochs = int(kwargs.get('epochs', 5))
+    patience = int(kwargs.get('patience', epochs))
+    conll_output = kwargs.get('conll_output', None)
+    span_type = kwargs.get('span_type', 'iob')
+    txts = kwargs.get('txts', None)
+    model_file = get_model_file('tagger', 'tf', kwargs.get('basedir'))
+    TRAIN_FLAG()
+
+    trainer = create_trainer(model_params, **kwargs)
+
+    do_early_stopping = bool(kwargs.get('do_early_stopping', True))
+    verbose = bool(kwargs.get('verbose', False))
+
+    best_metric = 0
+    if do_early_stopping:
+        early_stopping_metric = kwargs.get('early_stopping_metric', 'acc')
+        early_stopping_cmp, best_metric = get_metric_cmp(early_stopping_metric, kwargs.get('early_stopping_cmp'))
+        patience = kwargs.get('patience', epochs)
+        print('Doing early stopping on [%s] with patience [%d]' % (early_stopping_metric, patience))
+
+    reporting_fns = listify(kwargs.get('reporting', []))
+    print('reporting', reporting_fns)
+
+    last_improved = 0
+    for epoch in range(epochs):
+
+        trainer.train(ts, reporting_fns, dataset=False)
+        test_metrics = trainer.test(vs, reporting_fns, phase='Valid', dataset=False)
+
+        if do_early_stopping is False:
+            trainer.checkpoint()
+            trainer.model.save(model_file)
+
+        elif early_stopping_cmp(test_metrics[early_stopping_metric], best_metric):
+            last_improved = epoch
+            best_metric = test_metrics[early_stopping_metric]
+            print('New best %.3f' % best_metric)
+            trainer.checkpoint()
+            trainer.model.save(model_file)
+
+        elif (epoch - last_improved) > patience:
+            print('Stopping due to persistent failures to improve')
+            break
+
+    if do_early_stopping is True:
+        print('Best performance on %s: %.3f at epoch %d' % (early_stopping_metric, best_metric, last_improved))
+    if es is not None:
+
+        trainer.recover_last_checkpoint()
+        # What to do about overloading this??
+        evaluator = TaggerEvaluatorTf(trainer.model, span_type, verbose)
+        start = time.time()
+        test_metrics = evaluator.test(es, conll_output=conll_output, txts=txts, dataset=False)
+        duration = time.time() - start
+        for reporting in reporting_fns:
+            reporting(test_metrics, 0, 'Test')
+        trainer.log.debug({'phase': 'Test', 'time': duration})
